requirements.txt
# Brain Tumor Detection Project - Dependencies

# Deep Learning
tensorflow>=2.10.0

# Computer Vision
opencv-python>=4.7.0
opencv-python-headless>=4.7.0 # Required for Streamlit Cloud

# Data Processing
numpy>=1.23.0
pandas>=1.5.0
Pillow>=9.4.0

# Machine Learning
scikit-learn>=1.2.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Web App
streamlit>=1.28.0

# Jupyter Notebooks
jupyter>=1.0.0
ipykernel>=6.20.0

# Utilities
tqdm>=4.64.0
python-dotenv>=1.0.0
